Split LiveImapClient and ImapClientTests to fit SwiftLint file_length
The "No mailbox selected" fix from #358 grew both files past SwiftLint's
400-line file_length cap (LiveImapClient.swift hit 413, ImapClientTests.swift
hit 431). Lift the new code into sibling files:

- `LiveImapClient+TransportRetry.swift` — the actor-isolated extension
  carrying `withTransportRetry`, `isRecoverableTransportError`, and
  `isNoMailboxSelected`. Drops `withTransportRetry` from `private` to
  module-internal so the call sites in the main actor body still resolve.
- `ImapClientSelectCacheTests.swift` — the two #356 regression tests
  (`testFailedSelectResetsCacheSoNextSelectReissues`,
  `testNoMailboxSelectedTriggersReSelectAndRetry`) split into their own
  XCTestCase subclass so they can carry a private `signIn` helper without
  punching a hole in `ImapClientTests`'s encapsulation.

Pure refactor: no behavior change, no test change, no public API change.
